The global file now reads in data from a csv, and partitions it into dataframes corresponding to each pollutant, The UI has been updated and now is relevant to the project, and the server has had a huge edit.
The app is now centered on London, and can show any of the 5 pollutants by ticking the checkbox, the size of the spot corresponds to the value recorded in the csv, times an arbitary factor, which will need further investigation.

There is sometimes a shiny error, which completely stops the program from running through the server, but is totally fine with RStudio, if you restart the R instance.